TEP-IFPI-2017_8-master/studemy-api-master/requirements.txt
certifi==2018.11.29
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.3
django-cors-headers==2.4.0
django-debug-toolbar==1.10.1
django-filter==2.0.0
django-filters==0.2.1
djangorestframework==3.9.0
gunicorn==19.9.0
idna==2.7
Pillow==5.3.0
pkg-resources==0.0.0
psycopg2==2.7.6.1
python-decouple==3.1
python-firebase==1.2
pytz==2018.7
raven==6.9.0
requests==2.20.1
six==1.11.0
sqlparse==0.2.4
Unipath==1.1
urllib3==1.24.1
whitenoise==4.1.2
